+++ SharedLibs/Entities/Monsters/DummyMonster/GraphicDummyMonster.hpp
@@ -15,8 +15,8 @@ public:
 
 public:
     void Draw(sf::RenderTexture &rect) override;
-    vec2<int> GetRenderRect() override;
-    vec2<int> GetPosition() override;
+    vec2<float> GetRenderRect() override;
+    vec2<float> GetPosition() override;
 };
 
 #endif //R_TYPE_GRAPHICDUMMYMONSTER_HPP
